fix error messages and authorization for updating and deleting post and deleting comments

+++ backend/controllers/blogs.js
@@ -1,5 +1,4 @@
 import { prisma } from '../lib/prisma.js';
-import  jwt from 'jsonwebtoken';
 import "dotenv/config"
 
 async function createBlogPost(req,res){
@@ -14,6 +13,7 @@ async function createBlogPost(req,res){
          res.json('added')
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
@@ -21,9 +21,10 @@ async function createBlogPost(req,res){
 async function allBlogPosts(req,res){
     try{
         const posts = await prisma.posts.findMany();
-        res.send(posts)
+        res.json(posts)
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
@@ -34,9 +35,14 @@ async function blogById(req,res){
                 id : +req.params.blogId
             }
         })
-        res.send(post)
+
+         if(!post){
+            return res.status(404).json({error: 'Post not found'});
+        }
+        res.json(post)
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
@@ -54,6 +60,7 @@ async function updatePost(req,res) {
         res.json('edited')
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
@@ -67,21 +74,24 @@ async function deletePost(req,res) {
         res.send('deleted post')
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
 async function addComment(req,res){
     try{
-        await prisma.comment.create({
+             await prisma.comment.create({
             data:{
-                userId : +req.body.userId,
+                userId : req.user.sub,
                 comment: req.body.comment,
                 postId: +req.params.blogId
             }
         })
         res.json('comment added')
+        
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
@@ -95,19 +105,33 @@ async function allComments(req,res) {
         res.json(comments)
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     }
 }
 
 async function deleteComment(req,res) {
     try{
-        await prisma.comment.delete({
+        const comment = await prisma.comment.findUnique({
+            where: {id: +req.params.commentId}
+        })
+
+        if(!comment){
+            return res.status(404).json({error : "comment not found"})
+        }
+        if(req.user.role === "ADMIN" || req.user.sub === comment.userId){
+            await prisma.comment.delete({
             where:{
                 id : +req.params.commentId
             }
         })
         res.send('deleted comment')
+        }else{
+            res.status(403).json({error: "not allowed"})
+        }
+        
     }catch(error){
         console.error(error)
+         res.status(500).json({error: 'Server error'})
     } 
 }
 export{createBlogPost, allBlogPosts, blogById, updatePost, deletePost, allComments, addComment,deleteComment}
